Optimization and zoom bug fixing
Only render the bottom of the set, then reflect it to the top side. This improves performance and gets rid of asymetry due to having pixels being offset from the bottom (ex. the middle pixels of the set were not rendering the points with complex value of 0i, but something close to there.)
When zooming, temporarily get rid of the transformation matrix to draw a blank rectangle on the canvas. Before, if the set was dragged too fast, remnents would be left behind on the canvas element.